First pass at generated Version class.
Static version information that was hard-coded in the original 
source has been extracted and is now generated into a Version
class, using values from version.properties.  This should put
all version information into that properties file, so that it
only needs to be modified in a single place, and will propagate
all places where it is needed.

git-svn-id: https://svn2.assembla.com/svn/romraider/trunk@56 38686702-15cf-42e4-a595-3071df8bf5ea